requirements.txt
flask
ultralytics==8.3.179
opencv-python
pillow
numpy
pytesseract
werkzeug
